refactor: improve the folder structure for storybook component
