+++ src/temp/stories/landingPage/ExploreProject.stories.tsx
@@ -1,7 +1,8 @@
-import ExploreProject from '../Explore/ExploreProject';
+import ExploreProject from '../../Explore/ExploreProject';
 import type { Meta, StoryObj } from '@storybook/react';
 
 const meta: Meta<typeof ExploreProject> = {
+  title: 'Landing Page/ExploreProject',
   component: ExploreProject,
 };
