+++ src/temp/stories/projectDetails/ImagesSlider.stories.tsx
@@ -1,7 +1,8 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import ImagesSlider from '../ImagesSlider';
+import ImagesSlider from '../../ImagesSlider';
 
 const meta: Meta<typeof ImagesSlider> = {
+  title: 'Project Details/ImagesSlider',
   component: ImagesSlider,
 };
