+++ src/temp/stories/projectDetails/KeyInfo.stories.tsx
@@ -1,7 +1,8 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import KeyInfo from '../ProjectInfo/KeyInfo';
+import KeyInfo from '../../ProjectInfo/KeyInfo';
 
 const meta: Meta<typeof KeyInfo> = {
+  title: 'Project Details/KeyInfo',
   component: KeyInfo,
 };
