+++ src/temp/stories/projectDetails/MapPreview.stories.tsx
@@ -1,7 +1,8 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import MapPreview from '../MapPreview';
+import MapPreview from '../../MapPreview';
 
 const meta: Meta<typeof MapPreview> = {
+  title: 'Project Details/MapPreview',
   component: MapPreview,
 };
