+++ src/temp/stories/projectDetails/VideoPlayer.stories.tsx
@@ -1,6 +1,8 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import VideoPlayer from '../ProjectDetailVideoPlayer';
+import VideoPlayer from '../../ProjectDetailVideoPlayer';
+
 const meta: Meta<typeof VideoPlayer> = {
+  title: 'Project Details/VideoPlayer',
   component: VideoPlayer,
 };
